Melhorias na integridade DataBase

+++ src/main/java/dev_marcelo/maNotes/dto/DespesaDto.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor
 public class DespesaDto {
     private String nomeDaConta;
-    private Float valorDaConta;
+    private Double valorDaConta;
 }
